for TCP service scans, groups IPs by open port for improved efficiency | custom status updates for service scans | now accepts DNS names as targets/excludes | runs pre-check to ensure DNS resolves | UDP discovery scans now performed by masscan | greatly improved ext nmap host discovery command | greatly improved int host discovery by using nmap for <25000 hosts and masscan for >25000 hosts | custom status updates for nmap discovery scans | ext nmap host discovery command now customizable in ngineer mode | scan efficiency improvements: ext masscan port scan now uses nmap alives list as targets | new --rate option | new --session-list option | new --count-ports option | --count --> --count-hosts | new --dns-ip option | improved nmap status updates using --stats-every option | fixed nmap service scan resuming | skips UDP prompt if only TCP stage is set | includes network addresses (.0) in IP list and count and scans port 0 in open port scans | improved open ports extraction | improved excludes input | output file reorganization | banner outputs to stderr | UX improvements | required stage files improvements | other minor fixes and improvements

+++ installz0e.sh
@@ -25,12 +25,19 @@ done
 mkdir -p "$install_dir"
 
 #Copy Zero-E to the installation directory
-sudo cp ./zero-e "$install_dir/zeroe" && sudo chmod +x "$install_dir/zeroe"
-exitstatus=$?
+if [ ! -f ./zero-e ]; then
+  echo -e "\e[31m [X] Error: zero-e file not found in the current directory \e[0m"
+  exit 1
+else 
+  sudo cp ./zero-e "$install_dir/zeroe" && sudo chmod +x "$install_dir/zeroe"
+  exitstatus=$?
+fi
 
 if [ $exitstatus -eq 0 ]; then
     echo -e "\e[32m [+] Zero-E copied to $install_dir/zeroe \e[0m"
     echo -e "\e[36m [-] Zero-E can now be invoked as a command with \e[32mzeroe\e[36m \e[0m"
 else
     echo -e "\e[31m [X] Error: Failed to install Zero-E \e[0m"
+    echo -e "\e[31m     Ensure the zero-e file exists in your current working directory \e[0m"
+    echo -e "\e[31m     and that you have permission to write to $install_dir \e[0m"
 fi
